+++ no_usages/views_NO_USAGES.py
@@ -1,14 +1,16 @@
 import arcade
 import arcade.gui
 
+import gui.menu_view
+import gui.views.menu_view
 from gui.screen_configuration import ScreenConfiguration
 
 
 class Views:
 
     def __init__(self, screen_configuration: ScreenConfiguration):
-        self.game_view = main_game.GameView(screen_configuration)
-        self.menu_view = menu_window.MenuView(screen_configuration)
+        self.game_view = main_game.MainGameView(screen_configuration)
+        self.menu_view = gui.views.menu_view.MenuView(screen_configuration, view_controller)
